refactor(cli): extract ci cluster + hoist shared check symbols (5.0-alpha)
Fifth cut of the 5.0-alpha cli.ts god-module split. Extracts the `ci` command
and hoists the check/CI symbols it shares with cmdCheck + cmdSelfAudit (which
both stay in cli.ts) into a new neutral module.

New src/cli-checks-shared.ts (imported by cli.ts and commands/ci.ts, no cycle):
  - JsonCheck / JsonCheckOutput interfaces
  - maybeEmitCheckAttestation, autoInstallScanners (co-owned by cmdCheck)
  - CiFormat, detectCiFormat, emitGithubAnnotations, emitGitlabJunit
    (co-owned by cmdSelfAudit)
  - KIT_VERSION derived locally from package.json (no back-import of cli.ts)

New src/commands/ci.ts: cmdCi (exported), importing the shared symbols above.
  - Preserves process.argv.slice(2) (kit ci flags sit right after the command).
  - Dynamic import ./ci-init.js re-pathed to ../ci-init.js.

cli.ts: imports cmdCi + the shared symbols back; COMMANDS table unchanged;
prunes now-unused imports (gateStatus, evaluatePolicy — cmdCi-only). The
escapeWorkflowCmd/xmlEscape re-export for tests stays.

cli.ts: 4322 -> 3966 lines (-356; -2364 cumulative across PRs 1-5, from 6330 —
now under 4000). Typecheck, eslint (0 errors), full suite (2580), prettier all
green; kit ci (json/init) + self-audit smoke-tested.
